Support minimum 64bit integers

+++ src/writei.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
 
-void writeInteger(long n) {
-  printf("%ld", n);
+void writeInteger(long long n) {
+  printf("%lld", n);
 }
